feat: add IsDisabled property to tools and update views
Add the IsDisabled property to the Tool model to allow for 
soft deletion of tools. Update the HomeController to filter 
out disabled tools from the index view. Modify the Admin 
controller and views to support editing the IsDisabled 
property. Update the database schema with a new migration 
to include the IsDisabled column in the Tools table.